requirements.txt
salesforce-lavis
gradio
argostranslate